Reworked request recording
