+++ instrumentation/openai-client-instrumentation/src/test/java/co/elastic/otel/openai/EmbeddingsTest.java
@@ -30,7 +30,7 @@ class EmbeddingsTest {
   static final AgentInstrumentationExtension testing = AgentInstrumentationExtension.create();
 
   @RegisterExtension
-  static final OpenAIRecordingExtension openai = new OpenAIRecordingExtension("EmbeddingsTest.yml");
+  static final OpenAIRecordingExtension openai = new OpenAIRecordingExtension("EmbeddingsTest");
 
   @Test
   void basic() {
